backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
langchain==0.3.13
langchain-community==0.3.13
huggingface-hub==0.27.0
